pyproject.toml
[tool.poetry]
name = "saucedemo-05"
version = "0.1.0"
description = ""
authors = ["Igor Korobchenko <korobchenko.igor@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
selenium = "^4.6.0"
webdriver-manager = "^3.8.4"
black = "^22.10.0"
chromedriver = "^2.24.1"
pytest = "^7.2.0"
flake8-pyproject = "^1.1.0.post0"
poetry = "^1.2.2"
py-webdriver-manager = "^0.0.3"
conf = "^0.4.1"
pytest-html = "^3.2.0"
allure-behave = "^2.11.1"
allure-pytest = "^2.11.1"
testit-allure-adaptor = "^0.1.9"
allure-python-commons = "^2.11.1"
pytest-allure-spec-coverage = "^0.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
